# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API SMS ANGOLA",
        "version": "1.0"
    },
    "paths": {
...
        "/api/password/verifemail": {
            "post": {
                "tags": [
                    "Password"
                ],
                "description": "Vérification de l'email.",
                "operationId": "verifemail",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "lg",
                                    "email",
                                    "g_recaptcha_response"
                                ],
                                "properties": {
                                    "lg": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "fabio@yopmail.com"
                                    },
                                    "g_recaptcha_response": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vérification de l'email."
                    },
                    "400": {
                        "description": "Bad Request."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                }
            }
        },
        "/api/password/verifotp": {
            "post": {
                "tags": [
                    "Password"
                ],
                "description": "Vérification du code OTP.",
                "operationId": "verifotp",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "lg",
                                    "email",
                                    "otp"
                                ],
                                "properties": {
                                    "lg": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "otp": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vérification du Code OTP."
                    },
                    "400": {
                        "description": "Bad Request."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                }
            }
        },
        "/api/password/addpass": {
            "put": {
                "tags": [
                    "Password"
                ],
                "description": "Ajout de mot de passe.",
                "operationId": "addpass",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "lg",
                                    "email",
                                    "otp",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "lg": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "otp": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mot de passe modifié avec succès."
                    },
                    "400": {
                        "description": "Bad Request."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                }
            }
        },
        "/api/password/editpass": {
            "put": {
                "tags": [
                    "Password"
                ],
                "description": "Modification de mot de passe.",
                "operationId": "editpass",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "oldpass",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "oldpass": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mot de passe modifié avec succès."
                    },
                    "400": {
                        "description": "Bad Request."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
}